feat: Add toolbar and refactor
