+++ grandlite/__init__.py
@@ -1,19 +1,13 @@
 import argparse
-import datetime
 import pathlib
 import sys
 import tempfile
-from typing import Any, Protocol
 
 import networkx as nx
-import pandas as pd
 import requests
-from dotmotif import GrandIsoExecutor, Motif
-from grandcypher import GrandCypher
 from prompt_toolkit import HTML, PromptSession, print_formatted_text
 
-
-# If XML, assume GraphML
+from .prompts import ALL_PROMPTS, StatefulPrompt
 
 
 def _infer_graph_filetype_from_contents(filename: str) -> str:
@@ -87,209 +81,6 @@ def detect_and_load_graph(graph_uri: str) -> nx.Graph:
     return host_graph
 
 
-# Output and error types for the prompt:
-_Output = _Error = str | None
-# The response type for the prompt (railroad-style tuple)
-Response = tuple[_Output, _Error]
-
-
-class StatefulPrompt(Protocol):
-    """
-    A protocol that defines the interface for a stateful prompt.
-    """
-
-    def _get_state(self, graph_pointer: nx.Graph):
-        """
-        Retrieve the state associated with a graph pointer.
-        """
-        ...
-
-    def _set_state(self, state):
-        """
-        Set the state associated with a graph pointer.
-        """
-        ...
-
-    def prompt_text(self):
-        """
-        Return the prompt text.
-        """
-        ...
-
-    def query(self, input_text: str) -> Any:
-        """
-        Perform a single query using the given input.
-        """
-        ...
-
-    def submit_input(self, input_text: str) -> Response:
-        """
-        Submit the given input to the prompt.
-        """
-        ...
-
-    def prompt_kwargs(self) -> dict:
-        """
-        Return any keyword arguments that should be passed to the prompt.
-        """
-        ...
-
-
-class GrandCypherStatefulPrompt(StatefulPrompt):
-    """
-    A stateful prompt that uses Grand Cypher to query a graph using the Cypher
-    query language.
-
-    https://neo4j.com/docs/cypher-manual/current/
-
-    """
-
-    def __init__(self, graph_pointer: nx.Graph):
-        """
-        Initialize the prompt with a graph pointer.
-
-        Arguments:
-            graph_pointer: A NetworkX graph pointer.
-
-        """
-        self._graph = graph_pointer
-        self._last_results = None
-
-    def _get_state(self):
-        return {
-            "last_results": self._last_results,
-        }
-
-    def _set_state(self, state):
-        self._last_results = state.pop("last_results", None)
-        if len(state) > 0:
-            raise ValueError(f"Unknown state keys: {list(state.keys())}")
-
-    def prompt_text(self):
-        return "cypher> "
-
-    def prompt_kwargs(self) -> dict:
-        return {}
-
-    def query(self, input_text: str) -> Any:
-        results = GrandCypher(self._graph).run(input_text)
-        return pd.DataFrame(results)
-
-    def submit_input(self, input_text: str) -> Response:
-        if input_text.lower().startswith("save"):
-            if self._last_results is None:
-                return None, "No results to save."
-            args = input_text.split(" ")[1:]
-            if len(args) > 0:
-                format = args[0].split(".")[-1]
-                filename = args[0]
-            else:
-                format = "json"
-                iso = datetime.datetime.now().isoformat()
-                filename = f"results-{iso}.{format}"
-
-            if format == "csv":
-                self._last_results.to_csv(filename)
-            elif format == "jsonl":
-                self._last_results.to_json(filename, orient="records", lines=True)
-            elif format == "json":
-                self._last_results.to_json(filename, orient="records")
-            elif format in ["md", "markdown"]:
-                self._last_results.to_markdown(filename)
-            elif format == "html":
-                self._last_results.to_html(filename)
-            else:
-                return None, f"Unknown format: {format}"
-            return f"Saved results to {filename}.", None
-
-        try:
-            results = GrandCypher(self._graph).run(input_text)
-            self._last_results = pd.DataFrame(results)
-        except Exception as e:
-            return None, str(e)
-        return self._last_results.to_markdown(), None
-
-
-class DotMotifStatefulPrompt(StatefulPrompt):
-    """
-    A stateful prompt that uses grandiso to query a graph using the DotMotif
-    query language.
-
-    https://github.com/aplbrain/dotmotif
-    https://github.com/aplbrain/grandiso-networkx
-
-    """
-
-    def __init__(self, graph_pointer: nx.Graph):
-        self._graph = graph_pointer
-        self._last_results = None
-        self._first_prompt = True
-
-    def _get_state(self):
-        return {
-            "last_results": self._last_results,
-            "first_prompt": self._first_prompt,
-        }
-
-    def _set_state(self, state):
-        self._last_results = state.pop("last_results", None)
-        self._first_prompt = state.pop("first_prompt", True)
-        if len(state) > 0:
-            raise ValueError(f"Unknown state keys: {list(state.keys())}")
-
-    def prompt_text(self):
-        return ("(esc+enter to submit)\n" if self._first_prompt else "") + "dotmotif> "
-
-    def prompt_kwargs(self) -> dict:
-        return {"multiline": True}
-
-    def query(self, input_text: str) -> Any:
-        results = GrandIsoExecutor(graph=self._graph).find(Motif(input_text))
-        return pd.DataFrame(results)
-
-    def submit_input(self, input_text: str) -> Response:
-        self._first_prompt = False
-        if input_text.lower().startswith("save"):
-            if self._last_results is None:
-                return None, "No results to save."
-            args = input_text.split(" ")[1:]
-            if len(args) > 0:
-                format = args[0].split(".")[-1]
-                filename = args[0]
-            else:
-                format = "json"
-                iso = datetime.datetime.now().isoformat()
-                filename = f"results-{iso}.{format}"
-
-            if format == "csv":
-                self._last_results.to_csv(filename)
-            elif format == "jsonl":
-                self._last_results.to_json(filename, orient="records", lines=True)
-            elif format == "json":
-                self._last_results.to_json(filename, orient="records")
-            elif format in ["md", "markdown"]:
-                self._last_results.to_markdown(filename)
-            elif format == "html":
-                self._last_results.to_html(filename)
-            else:
-                return None, f"Unknown format: {format}"
-            return f"Saved results to {filename}.", None
-
-        try:
-            results = GrandIsoExecutor(graph=self._graph).find(Motif(input_text))
-            self._last_results = pd.DataFrame(results)
-        except Exception as e:
-            return None, str(e)
-        return self._last_results.to_markdown(), None
-
-
-# A mapping of query languages to their respective stateful prompts.
-_ALL_PROMPTS = {
-    "cypher": GrandCypherStatefulPrompt,
-    "dotmotif": DotMotifStatefulPrompt,
-}
-
-
 def prompt_loop_on_graph(host_graph: nx.Graph, query_language: str = "cypher"):
     """
     A prompt loop that allows the user to query a graph using a query language
@@ -301,20 +92,20 @@ def prompt_loop_on_graph(host_graph: nx.Graph, query_language: str = "cypher"):
             'cypher' and 'dotmotif'.
 
     """
-    session = PromptSession(
-        enable_history_search=True,
-    )
+    session = PromptSession(enable_history_search=True)
 
-    if query_language not in _ALL_PROMPTS:
+    if query_language not in ALL_PROMPTS:
         raise ValueError(f"No known query parser for language '{query_language}'.")
 
-    stateful_prompt: StatefulPrompt = _ALL_PROMPTS[query_language](host_graph)
+    stateful_prompt: StatefulPrompt = ALL_PROMPTS[query_language](host_graph)
 
     exiting = False
     while not exiting:
         try:
             text = session.prompt(
-                stateful_prompt.prompt_text(), **stateful_prompt.prompt_kwargs()
+                stateful_prompt.prompt_text(),
+                **stateful_prompt.prompt_kwargs(),
+                bottom_toolbar=stateful_prompt.bottom_toolbar(),
             )
         except KeyboardInterrupt:
             continue
@@ -367,6 +158,7 @@ def cli():
     )
     # Optional query parameter. If not provided, enters an interactive loop.
     argparser.add_argument(
+        "-q",
         "--query",
         help="If not provided, enters an interactive prompt.",
         default=None,
@@ -385,7 +177,7 @@ def cli():
 
     if args.query is not None:
         try:
-            results = _ALL_PROMPTS[language](host_graph).query(args.query)
+            results = ALL_PROMPTS[language](host_graph).query(args.query)
         except Exception as e:
             print(e)
             sys.exit(1)
